requirements.txt
###### Requirements with Version Specifiers ######
python-dotenv==1.1.1
requests==2.32.4